requirements.txt
absl-py==1.4.0
astunparse==1.6.3
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1674535352125/work
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
cPython==0.0.6
cycler==0.11.0
dataclasses==0.8
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dgl==1.0.1+cu113
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912934572/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.3.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
joblib==1.1.1
jsonschema==2.6.0
jupyter @ file:///home/conda/feedstock_root/build_artifacts/jupyter_1611871900595/work
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.5.0
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1610127569062/work
matplotlib==3.3.4
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1673904152039/work
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
networkx==2.5.1
notebook==4.4.1
numba==0.53.1
numpy==1.19.5
nvidia-dali-cuda110==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.1.5
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit==1.0.15
protobuf==3.19.6
psutil==5.9.4
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pymongo==4.1.1
pynndescent==0.5.10
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.7.1
PyYAML==6.0
pyzmq==19.0.2
qtconsole==4.4.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
simplegeneric==0.8.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.6
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1645693042223/work
threadpoolctl==3.1.0
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.0+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1631041982274/work
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.14
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.3.1
zipp==3.6.0